Add method to remove blocks
This feature is required to be exposed for allowing one to compile SCL sources automatically.
Silent mode is not available when compiling SCL sources, and a conflict (such as a pre-existing block) will spawn a graphical prompt and halt the process.
Therefore, it is needed to first delete pre-existing blocks