Add context class
Added class for encapsulating Simatic and Log handles.
The main reason for this is to provide some configurability to an otherwise fully static library.
The library itself makes no effort to manage the API handle or the log, but uses them freely.
It is up to a higher level of abstraction (namely a CLI or Server implementations) to provide these external resources and manage them.