Refactor API; Change getProgram
Changed getProgram to obtain an S7Program by its LogPath property, rather than trying to reach it by project > station > rack > parent module.
The idea is the same, but this way we don't need to know the rack name.
This brings a change in the API, as rack is replaced by the program name